--- src/model/Galaxia.java
@@ -1,35 +1,0 @@
-package model;
-
-public abstract class Galaxia {
-	public Galaxia() {
-
-	}
-
-	protected String nome;
-	protected String posicao;
-	protected double dimensao;
-
-	public String getNome() {
-		return nome;
-	}
-
-	public void setNome(String nome) {
-		this.nome = nome;
-	}
-
-	public String getPosicao() {
-		return posicao;
-	}
-
-	public void setPosicao(String posicao) {
-		this.posicao = posicao;
-	}
-
-	public double getDimensao() {
-		return dimensao;
-	}
-
-	public void setDimensao(double dimensao) {
-		this.dimensao = dimensao;
-	}
-}
